+++ src/components/Overlay.svelte
@@ -1,0 +1,3 @@
+<div class="fixed flex items-center justify-center top-0 left-0 right-0 bottom-0 z-50 bg-stone-700 bg-opacity-50">
+	<h1 class="text-white text-xl text-center">Things are happending! Please wait a moment, oke..?</h1>
+</div>