+++ src/routes/api/gdrive/poem/+server.js
@@ -1,9 +1,9 @@
-import { storePoem } from "$lib/server/gdrive";
+import { loadPoem } from '$lib/server/gdrive';
 import { json } from '@sveltejs/kit';
 
 /** @type {import('./$types').RequestHandler} */
 export async function POST({ request }) {
     const jsonRequest = await request.json()
-    const responseBody = await storePoem(jsonRequest)
+    const responseBody = await loadPoem(jsonRequest)
     return json(responseBody)
 }