+++ src/routes/export/+page.svelte
@@ -1,7 +1,5 @@
 <script>
     import { poemImage } from "../../stores/exported";
-
-    console.log($poemImage)
 </script>
 
 <h1 class="text-center pt-5 pb-5 text-lg">Here is the exported image of your poem. You can save or copy it~</h1>